Refactor translation cleanup logic: modularize utilities, patterns, and file scanning into separate files for improved maintainability. Replace inline functions with reusable imports.
